refactor: technical debt reduction — fix critical bugs, PHPStan cleanup, 114 Livewire tests, extract WordPressHttpClient
Critical bug fixes:
- Fix search using non-existent `domain` column (→ `url`) in Dashboard, Performance, Backups
- Replace blocking `dispatchSync` backup with async `dispatch` in Livewire HTTP context
- Fix Dropbox OOM: lower upload threshold from 150MB to 8MB (chunked upload for larger files)

PHPStan baseline reduction (169 → ~0 real errors):
- Fix ManifestService type hint: concrete class → WordPressApiServiceInterface
- Add startUploadSession/appendToUploadSession/finishUploadSession to DropboxDriver
- Add uploadToAbsolutePath + listFolders to StorageDriver interface + implementations
- Add @property annotations to 12 Models (UptimeCheck, UptimeMonitor, SecurityMonitor, etc.)
- Add type hints to closures/foreach in ~20 files (PluginManagerService, SitePlugins, etc.)
- Fix enum comparisons: BackupStatus, MonitorState used as strings → proper enum values
- Fix dead code: redundant comparisons, null coalesce on known arrays, always-true guards
- Add @method annotations for Eloquent scopes (RollbackPoint, StatusPageIncident, VulnerabilityAlert)
- Resolve DI: new ManifestService/RetentionService → app() container resolution

Design improvements:
- Extract WordPressHttpClient (292 LOC) from WordPressApiService (805 → 535 LOC)
- Move CONNECTOR_CHANGELOG from Livewire component to config/connector.php
- Remove 11 runtime Schema::hasTable guards (all tables exist in migrations)
- Add match default to SafeUpdateService, fix SecurityCategory enum as array key

Testing (1 → 17 files, 13 → 114 tests):
- New Livewire tests: SiteOverview, SiteBackups, SitePlugins, GeneralSettings,
  BackupsOverview, UptimeOverview, PerformanceOverview, DashboardSearch,
  CreateSiteWizard, Clients, SecurityDashboard, ProfileSettings,
  NotificationSettings, SiteReports, SecurityHardening, DataRetentionSettings

Documentation:
- Add backup/restore runbook (docs/runbooks/backup-restore.md)

+++ app/Helpers/FormatHelper.php
@@ -13,7 +13,7 @@ class FormatHelper
         }
 
         $units = ['B', 'KB', 'MB', 'GB', 'TB'];
-        $i = floor(log($bytes, 1024));
+        $i = (int) floor(log($bytes, 1024));
 
         return round($bytes / pow(1024, $i), $precision).' '.$units[$i];
     }
